Prefix when.js resource path for buster server/static
when.js is currently mounted at '/when.js' before being concatenated into
'/buster/bundle-0.6.js'. This can conflict with projects that have a file
in the rootPath named 'when.js' if it is exposed as a resource. In this
case, the project specific when.js is used instead of the when.js buster
intended.

Adding a '/buster' prefix to '/when.js' as the intended file is buster's
dependency and not the project's.