Don't ignore uncaught exceptions
While rewriting the browser-side wiring stuff, I inadvertantly moved the
window.onerror binding to after loading sources and tests. This is too
late to report on uncaught exceptions in those files outside the tests.
We now bind window.onerror immediately and queue exceptions until we
have a runner we can communicate the errors to.